Fix for AzureAD#136
The sub claim from the id_token should not be used as the plain text user name.
BUGBUG: this is just a temporary workaround of the real problem. When prompted by the ADFS server,
the user can use either DOMAIN\USERNAME or his/her email address as the account name, which does not
necessarily need to be the same with the userId parameter here. This possible inconsistency renders
the check extraneous and error-prone.