ADAL for Apple Watch with nasty interactive UI bits removed.
NOTE: in order to work with Apple Watch any native library was BOTH compile and target iphoneos and watchos. This is because the AppleWatch Extension is both on the Apple Watch and the iPhone. Therefore I simply used preprocessor commands to carve out the interactive bits with:

`#if !TARGET_OS_WATCH `

NOTE 2: Currently due to this above "runs in both iOS and Apple Watch" weirdness your PREPROCESSORS can do the wrong thing (see `AuthenticationContext.h`). For instance TARGET_OS_IPHONE is also set to 1 on watchOS, though TARGET_OS_IOS is 0. Anyone using TARGET_OS_IPHONE expecting a 0 since it's not on an iPhone will be crestfallen.