Fix callback cleanup
UIApplicationDidBecomeActive gets hit in more cases now, especially since iOS 9 and that's been screwing us up. Instead we're going to wait to get UIApplicationWillEnterForegroundNotification first before looking at UIApplicationDidBecomeActive, which narrows us down to only the case we care about. With that and the previous synchronization fix it appears that we shouldn't need to delay the cleanup anymore.