Initial Mac Check in
This will compile but not link. Need to add a couple more stub files for linking.

Ifdef'd out problematic code on iOS. I'll circle around to fix those later
Ported over NTLM UI, as it's relatively small and easy to do
Moved ADJwtHelper to the workplacejoin subfolder and it's closer to wpj related code