Broker Notifications and Resume Behavior (#482)
Add two more notifications for switching to broker and returning
Process and cache the broker result, even if we don't have a completion block to process it
Add an API to allow the developer to get the broker result from an interrupted sesions (e.g. from the app getting terminated) This will allow developers to mitigate low memory scenarios, especially on older devices without much memory.