Get 32-bit Mac Building
Most of the differences for the 32-bit build are because only the legacy ObjC runtime is supported. This eliminates support for a number of nice-to-have compiler features including ARC, declaring ivars in implementation files, auto-synthesization, and @synthesize expected the ivar to match the property name where in the modern runtime it'll pick the property name or the property name prefixed with a '_'. I still need to do the de-arcification manual memory management pass.

On top of that ARC can only be turned off on either a per-target or per-file basis, but not a per-architecture basis, so 32-bit build will have to be its own target and I'll lipo it into the 64-bit framework.