Build System Improvements
Hopefully this works, unfortunately there's a lot of special things about the setup of Travis (particularly around the lack of code signing certs) that's difficult to repro on my one local machine without ruining the dev environment and having to spend several hours rebuilding it.

In the past we weren't failing the build in all the cases we should have because error codes were being swallowed. That shouldn't happen anymore.

There's also a lot of improvements here in the loggging and folding, because now that we're building so many more things it's that much more difficult to pull out issues in the logs.