Rework the ADClientMetrics API.
We don't want to split setting data for different operations between the begin and end calls, because if multiple ADAL operations are happening at the same time the odds that we'll end up misreporting metrics between endpoints are fairly high. In this change we only set the metris data at the very end of the opertation, so in effect begin grabs any metrics data and clears it out, end sets the metrics data. If end gets hit multiple times before a begin we'll still end up dropping data though.