Use the same acquireToken: path for assertions.
In the before-times there wasn't an ADAuthenticationRequest object that neatly contained the request state. This caused the method parameters to multiply, and in some cases entire code paths would be duplicated to avoid having to add even more parameters to existing code paths. The only appreciable difference when using a SAML assertion is instead of going into the interactive webview auth we attempt to use the assertion instead. This way there's only one code path for traversing the cache and attempting to refresh tokens.