Output Single iOS Framework Distribution (#738)
* Explicitly set tab width to 4 spaces in the Xcode Project file

This will help new contributers align with the existing style

* Change strategy for handling application extension hosts to use runtime checks instead of separate framework distribution

The original strategy was to create a separate ADAL.framework output to link against for iOS application extension hosts. The problem with that approach is the consumer will end up with two different distributions of ADAL.framework installed on the device's `/Framework` directory. One would be built for the main app whereas the other is built for app extensions. This leads to an ambiguous state.

The new strategy is to distribute a single ADAL.framework and avoid calling non-application-extension safe APIs at runtime. This will work as long as it compiles (it does!) and passes app store review.

* Delete iOS application extension specific output targets and related code

We don't need this output anymore because the single ADAL.framework distribution now performs runtime checks to avoid non-application-extension safe APIs.

* Configure sample projects to embed a copy of `ADAL.xcodeproj`

This allows the sample project's `xcodeproj` to make `ADAL.framework` a dependency and later as a linked framework