Add back if check on RedirectUri error
In the broker refactor this if check was removed, however we didn't error out, so the code path still worked, we just got an extraneous (and erroneous) error message. This condition is checked very early on in the request so we fail reliably early, instead of late, which is why this wasn't caught, nor caused any run time issues. The erroneous log message though is obnooxious enough that it's worth fixing now.