Remove AUTH_NON_PROTOCOL_ERROR, return error directly
This behavior of slipping the error in the dictionary instead of returning it directly violates ObjC best practices, and results in some truly byzantine code paths where the code should be exiting early instead. In most cases this change results in no differences in code execution, just simplifies the code. In a couple cases it means that we exit a little earlier, but has no effect on the overall flow.