Remove error params from cache lock operations, throw exceptions instead
The only time we should ever see an error return from these lock operations is if the lock has been corrupted, in which case we’re already in a bad state with no reasonable way of recovering anyways. The app will crash shortly regardless and it is safer to exit quickly.