Use separate token cache for each thread
Occasionally (~1 in 1500 runs on my MBP, but much much more frequently on Travis) the first thread would finish fast enough that it’d write a RT out before the second thread read in a RT. This would cause the test to fail on the network mock. However because the network mock (for now) is hitting an NSAssert that caused an exception that would fail the test, but with no apparent reason in the log.