Reject protection space when no suitable cert is found
When Client Certificate or Device Authentication is
enabled and the client cannot find a suitable certificate to
satisfy the request, the protection space should be rejected
rather than allowing the default handler handle the request.
This allows additional protection spaces to be tried so that the
user can authenticate in another way.